moved conversation/TA example to it's own folder, added readme, package.json, and .env.example file
