+++ README.md
@@ -248,7 +248,7 @@ The [Concept Insights][concept_insights] has been deprecated, AlchemyLanguage's 
 
 ### Conversation
 
-Use the [Conversation](http://www.ibm.com/watson/developercloud/conversation.html) service to determine the intent of a message.
+Use the [Conversation][conversation] service to determine the intent of a message.
 
 Note: you must first create a workspace via Bluemix. See [the documentation](http://www.ibm.com/watson/developercloud/doc/conversation/overview.shtml) for details.
 
@@ -608,23 +608,11 @@ visual_recognition.classify(params, function(err, res) {
 ## Composing Services
 
 ### Integration of Tone Analyzer with Conversation
-Sample code for integrating [Tone Analyzer][tone_analyzer] and [Conversation][conversation] is provided in the [examples directory][examples]:
+Sample code for [integrating Tone Analyzer and Conversation][conversation_tone_analyzer_example] is provided in the [examples directory][examples].
 
-  * [examples/conversation_addons/tone_detection.js][tone_conversation_integration_example_tone_detection] - sample code to initialize a user object in the conversation payload's context (initUser), to call Tone Analyzer to retrieve tone for a user's input (invokeToneAsync), and to update tone in the user object in the conversation payload's context (updateUserTone).
-  
-  * [examples/tone_conversation_integration.v1.js][tone_conversation_integration_example] - sample code to use tone_detection.js to get and add tone to the payload and send a request to the Conversation Service's message endpoint. 
-
-Command to run the sample code
-
-``node examples/tone_conversation_integration.v1.js``
-
-Requirements to run the sample code
-
-  * [Tone Analyzer Service credentials][bluemix_tone_analyzer_service]
-  * [Conversation Service credentials][bluemix_conversation_service]
-  * [Workspace id][conversation_simple_workspace]
-  * credentials can be added directly to the [code][tone_conversation_integration_example] or to a .env file in the [examples directory][examples]
-
+## Integration Document Conversation with Retrieve and Rank
+See the [Document Conversion integration example][document_conversion_integration_example] about how to integrate the Document Conversion service
+with the Retrieve and Rank service.
 
 ## Unauthenticated requests
 By default, the library tries to use Basic Auth and will ask for `api_key` or `username` and `password` and send an `Authorization: Basic XXXXXXX`. You can avoid this by using:
@@ -674,10 +662,6 @@ See [CONTRIBUTING](https://github.com/watson-developer-cloud/node-sdk/blob/maste
 
 [conversation]: https://www.ibm.com/watson/developercloud/conversation.html
 
-[bluemix_conversation_service]: https://console.ng.bluemix.net/catalog/services/conversation/
-[bluemix_tone_analyzer_service]: https://console.ng.bluemix.net/catalog/services/tone-analyzer/
-[conversation_simple_workspace]: https://github.com/watson-developer-cloud/conversation-simple#workspace
-
 [personality_insights]: http://www.ibm.com/watson/developercloud/doc/personality-insights/
 [relationship_extraction]: http://www.ibm.com/watson/developercloud/doc/sireapi/
 [retrieve_and_rank]: http://watson.stage1.mybluemix.net/watson/developercloud/retrieve-rank.html
@@ -701,8 +685,9 @@ See [CONTRIBUTING](https://github.com/watson-developer-cloud/node-sdk/blob/maste
 [bluemix]: https://console.ng.bluemix.net
 [npm_link]: https://www.npmjs.com/package/watson-developer-cloud
 [request_github]: https://github.com/request/request
+
 [examples]: https://github.com/watson-developer-cloud/node-sdk/tree/master/examples
 [document_conversion_integration_example]: https://github.com/watson-developer-cloud/node-sdk/tree/master/examples/document_conversion_integration.v1.js
-[tone_conversation_integration_example]: https://github.com/watson-developer-cloud/node-sdk/tree/master/examples/tone_conversation_integration.v1.js
-[tone_conversation_integration_example_tone_detection]: https://github.com/watson-developer-cloud/node-sdk/tree/master/examples/conversation_addons/tone_detection.js
+[conversation_tone_analyzer_example]: https://github.com/watson-developer-cloud/node-sdk/tree/master/examples/conversation_tone_analyzer_integration
+
 [license]: http://www.apache.org/licenses/LICENSE-2.0
