+++ examples/conversation_tone_analyzer_integration/.env.example
@@ -1,0 +1,9 @@
+# see README.md for details
+
+CONVERSATION_USERNAME=<Conversation Username>
+CONVERSATION_PASSWORD=<Conversation Password>
+WORKSPACE_ID=<Workspace ID>
+
+TONE_ANALYZER_USERNAME=<Tone Analyzer Username>
+TONE_ANALYZER_PASSWORD=<Tone Analyzer Password>
+
