+++ examples/conversation_tone_analyzer_integration/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "conversation_tone_analyzer_integration",
+  "version": "1.0.0",
+  "description": "Example combining the Watson Conversation and Tone Analyzer services",
+  "main": "tone_conversation_integration.v1.js",
+  "author": "https://github.com/aprilwebster",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "bluebird": "^3.4.1",
+    "dotenv": "^2.0.0",
+    "watson-developer-cloud": "^2.0.3"
+  }
+}
