+++ examples/conversation_tone_analyzer_integration/tone_conversation_integration.v1.js
@@ -31,15 +31,14 @@
 /*eslint-env es6*/
 
 var watson = require('watson-developer-cloud');
-var tone_detection = require("./conversation_addons/tone_detection.js");
-require('dotenv').config({silent: true})
+var tone_detection = require("./tone_detection.js");
+require('dotenv').config({silent: true});
 
 
 /**
  * Instantiate the Watson Conversation Service
  */
 var conversation = watson.conversation({
-  url: 'https://gateway.watsonplatform.net/conversation/api',
   username: process.env.CONVERSATION_USERNAME || '<conversation_username>',
   password: process.env.CONVERSATION_PASSWORD || '<conversation_password>',
   version_date: '2016-07-11',
